Working on Enemy Line of sight
I accidentally committed without putting in a description so I will just
explain that one on this commit. I basically made it so that the enemy
is always calculating if the player is a certain distance away from the
it. I added a RANGE int in the global class, because I figured all
enemies would have the same general range. I now need to work on the
cone of sight.