fixed the enemy class
I changed up the enemy movement so that it was almost identical to the
player movement. I added the diagonal texture regions to the enemies,
and I also made it so that the enemies stopped in front of you, instead
of "in" you. I added the required arguments to the parameters of the
walls constructor because it generated that error after making Enemy
extend Body.